# use croaring serialization functions