@alexey-milovidov 2019-09-11
Do we need to mention backwards incompatibility (if someone was storing aggregation states on disk)?