@nicelulu 2019-09-16
However, older versions of the table column with bitmap will not work properly.